Update codegen utils (#213)

+++ csrc/algovivo/dynamics/backward_euler.h
@@ -28,7 +28,7 @@ void backward_euler_update_pos(
 
 extern "C"
 void backward_euler_update_vel(
-  float num_vertices, const float* pos0, const float* vel0, float* pos1, float* vel1, float h
+  int num_vertices, const float* pos0, const float* vel0, float* pos1, float* vel1, float h
 ) {
   const auto space_dim = 2;
   // vel1 = (pos1 - pos0) / h
